# tmp/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Check",
    "version": "1.0.0"
  },
  "paths": {
...
    "/auth/register": {
      "post": {
        "tags": [
          "Autenticación"
        ],
        "summary": "Registra un nuevo usuario",
        "description": "Endpoint público para crear una nueva cuenta de usuario.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserRegister"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Usuario creado exitosamente.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "example": "success"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "user": {
                          "$ref": "#/components/schemas/User"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Error de validación (ej. email duplicado).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JSendFail"
                }
              }
            }
          }
        }
      }
    },
    "/auth/login": {
      "post": {
        "tags": [
          "Autenticación"
        ],
        "summary": "Inicia sesión de un usuario",
        "description": "Autentica a un usuario con email y contraseña, y devuelve un token JWT.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserLogin"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Login exitoso, devuelve token.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "example": "success"
                    },
                    "data": {
                      "$ref": "#/components/schemas/AuthToken"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Credenciales inválidas.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JSendFail"
                }
              }
            }
          }
        }
      }
    },
    "/users": {
      "get": {
        "tags": [
          "Usuarios (Gestión)"
        ],
        "summary": "Obtiene la lista de todos los usuarios",
        "description": "Ruta protegida. Devuelve un array de todos los usuarios en la BD.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Lista de usuarios obtenida.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "example": "success"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "users": {
                          "type": "array",
                          "items": {
                            "$ref": "#/components/schemas/User"
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "No autorizado (token no válido o no provisto)."
          }
        }
      },
      "post": {
        "tags": [
          "Usuarios (Gestión)"
        ],
        "summary": "Crea un nuevo usuario (Ruta de Admin)",
        "description": "Ruta protegida. Permite a un admin crear un nuevo usuario (diferente de /auth/register).",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserRegister"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Usuario creado exitosamente.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "example": "success"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "user": {
                          "$ref": "#/components/schemas/User"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Error de validación."
          },
          "401": {
            "description": "No autorizado."
          }
        }
      }
    },
    "/users/{id}": {
      "get": {
        "tags": [
          "Usuarios (Gestión)"
        ],
        "summary": "Obtiene un usuario por su ID",
        "description": "Ruta protegida. Devuelve un solo usuario.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "description": "El ID numérico del usuario a buscar."
          }
        ],
        "responses": {
          "200": {
            "description": "Usuario encontrado.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "example": "success"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "user": {
                          "$ref": "#/components/schemas/User"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "No autorizado."
          },
          "404": {
            "description": "Usuario no encontrado."
          }
        }
      },
      "put": {
        "tags": [
          "Usuarios (Gestión)"
        ],
        "summary": "Actualiza un usuario",
        "description": "Ruta protegida. Actualiza el fullName o email de un usuario.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "description": "El ID numérico del usuario a actualizar."
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "fullName": {
                    "type": "string",
                    "example": "Joan Jose (Actualizado)"
                  },
                  "email": {
                    "type": "string",
                    "example": "joan.nuevo@ejemplo.com"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Usuario actualizado.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "example": "success"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "user": {
                          "$ref": "#/components/schemas/User"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "No autorizado."
          },
          "404": {
            "description": "Usuario no encontrado."
          }
        }
      },
      "delete": {
        "tags": [
          "Usuarios (Gestión)"
        ],
        "summary": "Elimina un usuario",
        "description": "Ruta protegida. Elimina un usuario de la base de datos.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "description": "El ID numérico del usuario a eliminar."
          }
        ],
        "responses": {
          "200": {
            "description": "Usuario eliminado.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "example": "success"
                    },
                    "data": {
                      "nullable": true,
                      "example": null
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "No autorizado."
          },
          "404": {
            "description": "Usuario no encontrado."
          }
        }
      }
    },
    "/categories": {
      "get": {
        "tags": [
          "Categorías"
        ],
        "summary": "Listar categorías",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Lista obtenida"
          }
        }
      },
      "post": {
        "tags": [
          "Categorías"
        ],
        "summary": "Crear categoría",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string"
                  },
                  "description": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Creado"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Token de autenticación JWT (Bearer). Introduce 'Bearer' [espacio] y luego tu token."
      }
    },
    "schemas": {
      "User": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "description": "El ID autoincremental del usuario.",
            "example": 1
          },
          "fullName": {
            "type": "string",
            "description": "El nombre completo del usuario.",
            "example": "Joan Jose Garcia"
          },
          "email": {
            "type": "string",
            "description": "El email único del usuario.",
            "example": "joan@ejemplo.com"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "description": "Fecha de creación."
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time",
            "description": "Fecha de última actualización."
          }
        }
      },
      "UserRegister": {
        "type": "object",
        "required": [
          "fullName",
          "email",
          "password"
        ],
        "properties": {
          "fullName": {
            "type": "string",
            "description": "El nombre completo del usuario.",
            "example": "Joan Jose Garcia"
          },
          "email": {
            "type": "string",
            "description": "El email único del usuario.",
            "example": "joan@ejemplo.com"
          },
          "password": {
            "type": "string",
            "format": "password",
            "description": "Contraseña (min 6 caracteres).",
            "example": "claveSegura123"
          }
        }
      },
      "UserLogin": {
        "type": "object",
        "required": [
          "email",
          "password"
        ],
        "properties": {
          "email": {
            "type": "string",
            "example": "joan@ejemplo.com"
          },
          "password": {
            "type": "string",
            "format": "password",
            "example": "claveSegura123"
          }
        }
      },
      "AuthToken": {
        "type": "object",
        "properties": {
          "token": {
            "type": "string",
            "description": "El token JWT para autenticación.",
            "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
          }
        }
      },
      "JSendFail": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "example": "fail"
          },
          "data": {
            "type": "object",
            "description": "Objeto o array con los errores de validación."
          }
        }
      },
...
    }
  },
...
}